Fix duration charts
