+++ src/frontend/components/user/UserVoteRow.tsx
@@ -1,4 +1,4 @@
-import { DSONANCE_COIN_SYMBOL } from "../../constants";
+import { DSONANCE_COIN_SYMBOL, MOBILE_MAX_WIDTH_QUERY } from "../../constants";
 import { fromNullable } from "@dfinity/utils";
 import { formatBalanceE8s } from "../../utils/conversions/token";
 
@@ -9,6 +9,7 @@ import ChevronUpIcon from "../icons/ChevronUpIcon";
 import ChevronDownIcon from "../icons/ChevronDownIcon";
 import DurationChart, { CHART_COLORS } from "../charts/DurationChart";
 import { map_timeline_hack } from "../../utils/timeline";
+import { useMediaQuery } from "react-responsive";
 
 interface Props {
   vote: SYesNoVote;
@@ -16,6 +17,8 @@ interface Props {
 }
 
 const UserVoteRow = ({ vote, selected }: Props) => {
+
+  const isMobile = useMediaQuery({ query: MOBILE_MAX_WIDTH_QUERY });
 
   const { data: debt_info, call: refreshDebtInfo } = protocolActor.useQueryCall({
     functionName: "get_debt_info",
@@ -34,7 +37,7 @@ const UserVoteRow = ({ vote, selected }: Props) => {
   );
 
   return (
-    <div className={`rounded-lg p-2 shadow-sm bg-slate-200 dark:bg-gray-800 hover:cursor-pointer w-full ${ selected ? "border-2 dark:border-gray-500 border-gray-500" : "border dark:border-gray-700 border-gray-300"}`}>
+    <div className={`flex flex-col items-center rounded-lg p-2 shadow-sm bg-slate-200 dark:bg-gray-800 hover:cursor-pointer w-full ${ selected ? "border-2 dark:border-gray-500 border-gray-500" : "border dark:border-gray-700 border-gray-300"}`}>
       <div className={`grid grid-cols-[minmax(24px,auto)_minmax(100px,1fr)_minmax(60px,auto)] gap-2 sm:gap-x-8 w-full items-center px-2 sm:px-3`}>
 
         <div className="flex justify-self-start">
@@ -56,20 +59,22 @@ const UserVoteRow = ({ vote, selected }: Props) => {
 
       {
         mined_dsn && selected &&
-          <DurationChart
-            duration_timelines={new Map([
-              [
-                "earned",
-                { timeline: map_timeline_hack(mined_dsn.amount, (contribution) => contribution.earned), color: CHART_COLORS.BLUE },
-              ],
-              [
-                "pending",
-                { timeline: map_timeline_hack(mined_dsn.amount, (contribution) => contribution.pending), color: CHART_COLORS.PURPLE },
-              ],
-            ])}
-            format_value={(value: number) => formatBalanceE8s(BigInt(value), DSONANCE_COIN_SYMBOL, 2)}
-            fillArea={true}
-          />
+          <div className={`w-2/3 ${isMobile ? "h-[200px]" : "h-[300px]"}`}>
+            <DurationChart
+              duration_timelines={new Map([
+                [
+                  "earned",
+                  { timeline: map_timeline_hack(mined_dsn.amount, (contribution) => contribution.earned), color: CHART_COLORS.BLUE },
+                ],
+                [
+                  "pending",
+                  { timeline: map_timeline_hack(mined_dsn.amount, (contribution) => contribution.pending), color: CHART_COLORS.PURPLE },
+                ],
+              ])}
+              format_value={(value: number) => formatBalanceE8s(BigInt(value), DSONANCE_COIN_SYMBOL, 2)}
+              fillArea={true}
+            />
+          </div>
         }
 
     </div>
